create-kernel-tasks: Handle multiple primary kernels per series
The way tasks were handled internally was written under the assumption
that each series will only have one source that is the source for
derivatives or backports.

This started to be wrong when focal/linux-oem-5.6 came along. And now
that we plan to support no longer series in backports (like 5.0 based
kernels in bionic) by making the related hwe kernel a new primary for
those kernel versions, this needs fixing.

Attempt to do so by moving from a series name as the task key to a
handle format (<series>:<source>).

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kleber Souza <kleber.souza@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>