sbin/logrotate: logrotate helper
Add a simplified logrotate helper.  This uses the .*.logrotate
configuration fragments in the specified directory to rotate the contained
logs.  If no fragment is supplied for a logfile (*.log) in the directory
one is created using the defaults.  This allows an application to supply
logrotate configuration such as post rotate commands by pre-creating
this configuration file.

Signed-off-by: Andy Whitcroft <apw@canonical.com>