cranky: Update cranky close helper
Consider a tree a rebase tree if $DEBIAN/etc/update.conf exists and
also contains a direcory which DEBIAN_MASTER defines.
With multiple primaries there will be primary kernels which are not
using debian.master and we expect to have more special trees like
uc20-efi.
Only insert the parent changelog delta if the current tree is a rebase
tree.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>