swm-publishing: limit publication scanning to the oldest 2000 entries
We are blowing our time budget when we have a lot of entries and get
killed of making the issue worse.  The publication iterators are cheap
enough to take slices from so handle the bottom 2000 entries only
when we see more than this.  We will handle more of the backlog when
we next scan until we catch up.  At 2000 per 5m run we will catch up
something like the haskall migration (at 58000 publications) in a couple
of hours.  Not ideal but better than just dying in a heap.

Signed-off-by: Andy Whitcroft <apw@canonical.com>