create-kernel-tasks: Change source selection
Initially, when specifiying which source to use (--source), the series
given had to be the series of which the source was derived. This was the
result of how tasks were aranged by series and the mental model of
considering only the top-down model used when selecting whole series.

But this would now fail when I experimented with creating
bionic:linux-hwe which is now a source in bionic. Using the arguments
'--series bionic --source linux-hwe' would try to add a new task for
xenial:linux-hwe as well (since it is a backport from bionic). And no
way to work around this.

This fix this, I propose to switch the meaning of --series when used
together with --source. It would now mean a source of that name in the
given series only. Using --series alone still means anything that is a
derivative of the given series (like before).

Doing that it turned out that I can simplify a lot of places where I
initially looped over all provided series to pick tasks. That check was
already made before when collecting the list of tasks.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>