announce: announcer-ckct -- message announcer
Pick up messages of interest flowing within the Ckct testing system
and announce them into the announcer.  Supports connection to the "new"
and optionially the "old" Ckct environments.

We spawn one thread connected to the announcer, we listen on this
connection for command messages (currently the quit command), and out
going messages are piggy-backed on the same connection.  We spawn a
second thread connected to the Ckct exchange on the new rabbitmq server.
We optionally spawn a third thread connected to the Ckct exchange on
the old rabbitmq server.  We bind for announce worthy messages on each,
format them and send them to the announcer.

Signed-off-by: Andy Whitcroft <apw@canonical.com>