cranky open: Relax the search for the previous ABI
With --reuse-abi, the previous kernel version needs to match the current
version but that's not necessarily true. When creating a new kernel tree,
the debian.<foo>/* files are typically imported from an older already
existing kernel which might be based on a different kernel version. Relax
the search for the previous ABI directory to handle that case.

Signed-off-by: Juerg Haefliger <juergh@canonical.com>
Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>