swm: bug -- maintenance record are not crank iteration transient
Maintenance records are generated when we scan the build status.  This is
cached and so if we reset the maintenance records on each iteration we
will always reset them away.  They are already initialised empty when
the tracker is made initially so they are crank transient rather than
crank-iteration transient.  Drop the errant reset.

Signed-off-by: Andy Whitcroft <apw@canonical.com>