info/kernel-series: Add -hwe-20.04 and -hwe-20.04-edge variants to groovy:linux
Starting in focal the hwe kernels should roll users forward to
the next hwe kernel on upgrade rather than parking them on the
generic kernel. We will use variants in the interim releases
to support upgrades, and then in 22.04 transitional packages
will be used to move users to the 22.04 hwe packages.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>