swm: swm-additional-checks -- initial stab at an exceptions helper
We treat some trackers differently at cirtain times.  For example when when
roll the HWE kernel from one major kernel version to another we need to
do additional upgrade testing.  We have previously added a kernel-signoff
task to represent that additional work needs to be done and to block the
trackers promotion until it is complete.

Create swm-additional-checks an (at least initially) manual step run
against the instantiated trackers which detects any such situations and
take appropriate action.  Initially detecting when a meta package rolls
from one major version to another.  This is detected regardless of whether
it bumps within a package or moves from one to another as happens with
our variant supporting kernels.

Signed-off-by: Andy Whitcroft <apw@canonical.com>