adt-matrix: ADT_TEST_TRIGGERS parsing issue
We are attempting to acertain the kernel package from any of the
ADT_TEST_TRIGGERS.  We have code attempting to connect results
for the main package itself back to the meta, but this does not
actually make sense as we are 1) not running the kernel if we don't also
have a meta trigger and 2) the code checks the wrong packages list and
so matches all test packages too.  Drop this errant check.

Signed-off-by: Andy Whitcroft <apw@canonical.com>