adt-matrix: is_proposed check is bogus
We are meant to classify any test which is done for a test package with
the version in proposed separatly (currently they are ignored).  We try
and do this by checking if there is a trigger for tested package unless
it is one of our source packages.  However we do this against the test
package list (we do not actually have a source package list).

Use KernelSeries to find the list of source packages for this kernel and
check against that.

Signed-off-by: Andy Whitcroft <apw@canonical.com>