adt-matrix: record results against proposed and report them in the matrix
Due to the errant -proposed check we have tended to mix the -proposed and
-updates versions together highest version winning.  If someone triggered
a test package -proposed version we would move miss immediatly there.
We are actually interested in test results against both the -updates
version in the sense that the results against that pocket should represent
the actual status for this package should it migrate alone; hinting can
be used against this.  The -proposed versions are interesting when we have
tried to fix the package and want to see how it would be if it migrated.

Now that we are correctly detecting that these tests are against -proposed
post them into a separate result structure (they are posted against arch +
'+' currently) and use this structure to add an optional second result
for the latest version only.

Signed-off-by: Andy Whitcroft <apw@canonical.com>