cranky: generate-update-conf: Add leading ~ or + to BACKPORT_SUFFIX
BugLink: https://bugs.launchpad.net/bugs/1902957

We currently rely on our packaging to add a ~ character in front
of BACKPORT_SUFFIX in version strings. However we have started
doing forward ports of kernels to the development series as well,
where we want this leading character to be + rather than ~. It's
difficult to figure out which leader is appropriate from our
kernel packaging, whereas when generating update.conf we have the
kernel series info which makes it rather trivial.

Update generate-update-conf to add the leader to BACKPORT_SUFFIX.
This will be a + if the "backport" series is later than the
source series, otherwise it will be a ~.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>