cranky: Add list-derivatives command
This is hopefully helping to keep on top of the ever growing list of
kernels we do. Given one or more sources in handle format, it will
produce a list of kernels that relate to the initial list (and the
kernel sources that were given).
This should be useful to populate for example the list of affected
kernels when one does a re-spin.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andrea Righi <andrea.righi@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>