swm-publishing: limit getPublished{Sources,Binaries} to linux- prefix
When scanning an open archive with a full source/binary scan limit searches
to those items which are likely kernel packages.  For sources all of
them are named linux-*.  For binaries all publications either contain
linux-libc-dev or linux-image-* and we are only interested in whether we
have any publication in an architecture as all publications are atomic.

Signed-off-by: Andy Whitcroft <apw@canonical.com>