devel: Add script to create UEFI vms for secureboot testing
Add devel/mk-uefi-vm, which can be used to create UEFI virtual
machines for amd64 and arm64, with secure boot enabled by default
on amd64. Cloud images are used to simplify and speed up
installation, with ssh access automatically enabled with the
user's key (or one specified on the command line).

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>