swm: package -- delay instantiation of build information
We currently instantiate all routes and all builds in those routes on
creation of the WorkflowBug which is horribly expensive.  Replace the
static data with a delayed instantiation object which only scans the
build when we request one of the attributes.

For trackers without versions in particular this saves a huge amount
of scan time.

Signed-off-by: Andy Whitcroft <apw@canonical.com>