create-kernel-tasks: Support handles and virtual primaries
This change does two things:
- Add a --handle selector for sources
- Tries to handle kernels which base on a devel source but
  practically are in a way stable

For the new --handle <series>:<source> format, the series of the first
handle that is provided decides whether this is treated as development
or stable cycle (no cards and any date for devel). And there is no
automatic selection other than snap tasks. There is just no sane way
to decide which kernels to include automatically and which not. On the
plus side, there is no restriction to a single series.
Using the old forms of selection should still be possible and yielding
the same results as before.

On the new handling of certain kernels based on development. Right now
this is more or less linux-oem-5.10. Although it is based on the 5.10
development kernel, it is treated as a Focal OEM kernel and as such
falls logically more into the stable cycle. It was possible to select
it with --series focal --source linux-oem-5.10 but logic then says
that if it is a derivative, then the parent kernel must also exist in
the same cycle. However development and stable cycles are grouped in
different name spaces. The only half-sane way of handling this seems
to be treating a source as primary if the parent is in development
series. Which is what was done.
[PS: of course now that this was added, the oem-5.10 kernel was re-
     defined as primary. Only good thing is that that part is rather
     small and I would leave the change in for next time...]

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>