swm: package -- handle duplicate Publication records
When a publication moves from Pending to Published we also schedule the
previous Publication records to move to Superseded.  This does not
however happen in a single transaction so it is possible to see two
records in Published.  Request the records in date order and take the
newest one (the first).

Signed-off-by: Andy Whitcroft <apw@canonical.com>