swm: bug,wft/workflow -- ensure we update is_closed et al on closure
We are not dropping the -live tags from trackers because we do not
re-evaluate is_closed at al after changing the workflow task status.

As things stand check_is_valid has a strange API, it is a class member
but is passed s.bug and returns the various status flags which then must
be assigned.  Switch this function to a proper class function picking
s.bug up itself and setting the various state flags directly.

Now that check_is_valid is reusable call it to update these state flags
when we manipulate the workflow task status.

Signed-off-by: Andy Whitcroft <apw@canonical.com>