swm: packate,wft/promote_to_* -- track packages to promote
When we move a promotion task to 'Confirmed' determine which packages
are missing from the destination.  Record that list as a promotion hint
in "delta" indexed by the taskname.  Recording them by taskname allows
us to better follow the history of the promotions and in principle lets
us have multiple such operations in play at once (such as might happen
with promote-to-proposed and promote-signing-to-proposed with an
embargoed package such as lrg).

Signed-off-by: Andy Whitcroft <apw@canonical.com>