swm: wft/prepare_package -- detect early close of prepare-package-*
If a prepare-package task is closed early manually or (more commonly) if
the package is replaced in the build route without informing swm (by
updating the related version) the prepare-package can not be considered
complete.  Start monitoring the builds when in Fix Released and should
they no-longer be present pull the task back to In Progress.  This will
trigger missing tag and missing upload reports in the dashboard as
appropriate.

Signed-off-by: Andy Whitcroft <apw@canonical.com>