ktl/workflow: add default assignee for kernel-signoff
When we originally added kernel-signoff it was an entirely manual task.
This means a default assignee had no meaning so it was not added.  However,
when we reset a tracker using tb-task-reset we use the presence of a
default assignee to validate the task.  Therefore add a default assignee
of canonical-kernel-team.

Signed-off-by: Andy Whitcroft <apw@canonical.com>