swm: bug,package -- separatly track the version of the source
It is important that we revisit the cached package versions if the
primary version on the tracker changes.  Currently we check that against
the main package version as those are they same.  However this breaks
down if we do a partial respin, such as a meta-only or LRM-only respin,
as we may not have a main package.  Add a new `source` entry to the
versions table which tracks the title version and use it for title
validation.

Signed-off-by: Andy Whitcroft <apw@canonical.com>