swm: package -- if source is invalid we cannot check test flavours
The data needed to generate the meta mapping is stored in kernel-series.
If we have no valid kernel-series source pointer we will crater trying
to generate it.  Return a null list.  We will catch the bad source
linkage elsewhere and put the whole tracker into a crank-failure so it
will never be consumed in anger.

Signed-off-by: Andy Whitcroft <apw@canonical.com>