mainline-builds: cod-mainline-core -- order possible tips by commit date
For unstable we add a master-<base-version> tip to the mix.  If this tip
is out of date we just make a huge mess of the world.  Switch to using
a reverse commitdate ordered list of tips.  This will both elide
absent tips from the list and ensure we use the newest available in all
cases.  For unstable we will make a more sensible decision, for
primaries in stable series we should never see the order invert as we
never push new code to master anyway.

Signed-off-by: Andy Whitcroft <apw@canonical.com>