mainline-builds: cod-mainline-broken -- handle core forms
We now have a core job which is separatly manged from the architecture
test jobs.  When finding an issue in a test job it may be appropriate
to resubmit just that architecture job (for example out of disk space
on a specific builder), or the whole core job (for example bad config
selection).  Add support for separating core from architecture jobs and
marking failures as either affecting core or leaf.

Signed-off-by: Andy Whitcroft <apw@canonical.com>