adt-matrix: adt-seed-accumulate -- use kernel-series to acertain package hierachy
Now that kernel-series has the full and accurate package set and hierachy
for a source we no longer should use the dependencies between packages
reported by britney as a hint.  This can be missing packages if consumed
at the wrong time and it can also (recently) have implicit dependencies
which are simply wrong for our use case.

Signed-off-by: Andy Whitcroft <apw@canonical.com>