swm: add support for adjunct build PPAs and direct ancillary packages lrs/lrg to them
We are introducing a private adjunct PPA for each build PPA, this is used
for components of an otherwise public build which must remain embargoed.
We also introduce the concept of an ancillary package, one which is
entirely derived from another.  These are version locked together and the
ancillary packages themselves have no source repo, tags or independant
versions.  Finally drive ancillary packages for lrm via the adjunct
PPAs.

Signed-off-by: Andy Whitcroft <apw@canonical.com>