swm: wft/verification_testing -- switch to scanning bug status internally
sru_report is not able to scan private trackers.  Now that we are able
to get the changes file, use the list of bugs expressed in there as
our driver for status.  Grab the bug list from the changes file from
our package subtracting any listed on our master-bug (when present).
Scan the remaining list using our authenticated launchpad connection so
we can see the private ones too.  Accumulate an overall status and use
this to drive the verification-testing task.

Signed-off-by: Andy Whitcroft <apw@canonical.com>