swm: package -- older_tracker_in_ppa needs to consider first run specially
When we are a new tracker, there is no existing entry for us.  This
prevents the "older" tracker algorithm detecting we are the oldest (as
we are not present).  Additionally check if the current tracker has an
newer cycle number, if we are newer we should be before (and arn't yet).

Signed-off-by: Andy Whitcroft <apw@canonical.com>