swm: bug -- a bug is not a replaces candidate if in updates
When a tracker has made it as far as being copied to -updates it is not
a candiate for duplication.  It is not going to be in the way (now or
very shortly) as all packages will be in -updates and we mostly care
about the build -> -proposed transition for duplication.

Drop some redundant parentheses on status comparisons which are benign
but make the right hand sides look like tuples when they are not.  They
add no syntactic value and lead to viewer confusion.

Signed-off-by: Andy Whitcroft <apw@canonical.com>