swm-publishing: instantiate our PPA access key
The act of giving PPA access to a user (such as the swm bot) does not
actually allocate a key.  The key is only instantiated on the first
viewing of the subscription.  This occurs when we view the subscription
on a personal +archivesubscriptions page via the
getArchiveSubscriptionURL() call.

Note that we must work around a python3 launchpad lib redirect handling
issue when using the lp.me link.  The HTTP spec says that any redirect
must be followed using GET which breaks many of the API calls when
applied to the lp.me which is a redirect to the logged in user.  This
link must be manually followed to avoid this.

Signed-off-by: Andy Whitcroft <apw@canonical.com>