swm: bug -- fix fallback swm-properties dumper
When saving swm-property data into the tracker description we
intentionally wrap it in order to avoid launchpad wrapping in the text
widget.  There is a small risk that we will break the yaml formatting
when we do this.  To avoid this we format the props dictionary into a
string and convert it back.  If the before and and after data is not
identicle we reconvert it without wrapping.

We have finally generated an swm-properties field which did not have the
same meaning once wrapped and fell back to the unwrapped dumper;
cratering in the process.  Fix this up.

Signed-off-by: Andy Whitcroft <apw@canonical.com>