adt-tools: adt-submit-bug: Improve bug description
When we generate the bug reports from ADT failures it seems that the
terse text causes people to assume we want to shift blame to the
tested package.
To avoid that, I have added a boilerplate text that explains things
a little. Also I changed the bug creation to add a kernel task in
parallel to the package task if the tested package was not the kernel
itself.
Last but not least, I added a usage message if one runs the command
without arguments. So people do not have to check the code each time
they run this.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>