swm: package,wft/promote-to-proposed -- search further for released packages
When trying to determine if it is safe to promote a package (that the
package which will be replaced is fully released) we need to continue
through the signing route as that is routinely cleaned after the signed
artifacts are copied out to proposed.  Make the "after" pocket a list
and add Proposed as an intermediate when scanning.

Signed-off-by: Andy Whitcroft <apw@canonical.com>