swm: package -- pocket clear handle embargoed packages
When searching later pockets to confirm that the specified pocket does
not contain live kernel packages we need to take account of packages
which are embargoed and therefore do not proceed beyond a cirtain
pocket.  If everything other than the embargoed packages have moved
beyond the specified pocket then the embargoed package has performed its
function and can be replaced.  Therefore for an embargoed package
consider it present in a pocket for which it is not permitted.

Signed-off-by: Andy Whitcroft <apw@canonical.com>