swm: wft/syn_prepare_packages -- indicate when all builds are complete
When all builds are complete but the prepare-packages are not complete we
will just say "building in <pocket>" which is confusing as we are not in
fact building.  For example if a package tag is missing we will hold the
`prepare-package` task open even after the build itself has completed
in PPA.  Add "(all builds complete)" in this scenario.

Signed-off-by: Andy Whitcroft <apw@canonical.com>