info/kernel-series.yaml: Move 16.04/Xenial into ESM support
This marks up the 16.04/Xenial series as ESM. Because there is no
support for that mode for snaps, it has been decided to just continue to
provide those publicly for free. The remaining question for snaps is if
and how they would be tested.
For linux, linux-fips, and linux-hwe supported arches are reduced to
amd64, arm64, and s390x.
The cloud kernels remain supported under ESM (including linux-kvm) but
linux-deeplens, linux-raspi2, and linux-snapdragon are no longer
supported.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>