info/kernel-series: re-convert focal:linux-intel to a non-backport kernel
Since we are dropping all architectures except for amd64 it becomes
quite complex to manage the re-sync of configs and rules from
debian.master to debian.intel that are happening for backport kernels.

So, re-convert linux-intel to a non-backport kernel for now. Ideally we
should have a cloud kernel hirsute:linux-intel and then we should
backport that to focal, but we only have focal:linux-intel based on the
hirsute kernel (5.11), so we need to workaround the short-circuit in
this way to reduce the dirtiness.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>