swm: package: bugs -- skip pockets which will not give us the changes file
We pick the changes file from the first pocket we find the package.  But
when the kernel is in proposed and the PPA version has been replaced by
a later spin we may find the package in signing.  This PPA is private
and will refuse to offer up the changes file.  Handle the permissions
error this produces and move on to the next pocket.

Signed-off-by: Andy Whitcroft <apw@canonical.com>