swm: snap -- ESM snaps are on canonical-kernel-esm
Launchpad limits where a snap object can be if it is going to access
private objects; it must be on the same team as those private objects.
With uc16 now in ESM its kernels have moved to the ESM project.  Expect
the snaps for kernels in ESM to be there too.

Signed-off-by: Andy Whitcroft <apw@canonical.com>