cranky/open: Drop useless ABI check and add an exit handler
Towards the end of the script, we currently check if the expected ABI
directory exists and emit a warning if it doesn't. I seriously doubt that
anybody pays attention to that warning so either fail hard or drop the
check altogether. I choose the latter.

While at it, add an exit handler that prints an error if the script
terminates with a non-0 exit status and also move the functions hl() and
run() to the begining of the script to group all functions together.

Signed-off-by: Juerg Haefliger <juergh@canonical.com>
Acked-by: Andrea Righi <andrea.righi@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>