devel: mk-uefi-vm -- update OVMF secureboot template path
In impish my VMs started hanging early in boot, either during the EFI
stub or shortly thereafter. I found that the path used for the OVMF
secureboot template is a symlink. Switching the path in the script to
the target of the symlink gets things moving, so let's just use that.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>