swm: package -- feeder completed time need a little wriggle room
When we copy to signing we will publish main and the source for signed
at the same time.  The signed will build and may fail because we have
not yet built and published the signing components.  We therefore need
to consider further retries long enough to see the next publisher run
complete.  This is normally done within an hour, so allow retries up to
two hours after.

Signed-off-by: Andy Whitcroft <apw@canonical.com>