adt-matrix: add adt-remove-package-version script to fix caches
This script allows to fix up cached adt-matrix results when packages
are removed from the archive. For example, glibc ubuntu9.3 sru got
remove from focal-proposed. When that happens cached results against
ubuntu9.3 should be dropped, and the latest known version of glibc
should be downgraded to ubuntu9.2

The script's --help in detail expalains when and how to use it.

This script is rewritten as dual python2 or python3 script. adt-matrix
currently runs as python2, and there are differences in the serialized
json ouptut between python2 and python3. To keep the updated cache
diffable this script should be executed using the same python as the
adt-matrix itself.

Here is a sample change that it can generate:

    $ adt-remove-package-version --series-cache ./focal.cache --package glibc --remove-version 2.31-0ubuntu9.3 --set-version 2.31-0ubuntu9.2
    $ diff ../../kernel/focal.cache focal.cache
    @@ -264632,14 +264632,6 @@
	       "20210427_150453_a9e3b@"
	     ]
	   },
    -      "glibc": {
    -        "amd64": [
    -          true,
    -          "2.31-0ubuntu9.3",
    -          "https://autopkgtest.ubuntu.com/results/autopkgtest-focal/focal/amd64/g/glibc/20210427_194311_ef12a@/",
    -          "20210427_194311_ef12a@"
    -        ]
    -      },
	   "nvidia-graphics-drivers-418-server": {
	     "amd64": [
	       true,
    @@ -297083,7 +297075,7 @@
	 "sysdig s390x": "0.26.4-1ubuntu0.3",
	 "apparmor s390x": "2.13.3-7ubuntu5.1",
	 "langford ppc64el": "0.0.20130228-6build1",
    -    "glibc amd64": "2.31-0ubuntu9.3",
    +    "glibc amd64": "2.31-0ubuntu9.2",
	 "jool ppc64el": "4.0.7-1",
	 "glibc ppc64el": "2.31-0ubuntu9.2",
	 "xtables-addons s390x -proposed": "3.9-1ubuntu0.2~20.04.1"

This tool should help with fixing the currently MISS glibc results in
the current SRU cycle's adt-matrix for focal.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>
Acked-by: Seth Forshee <seth.forshee@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>