cranky: chroot: install dctrl-tools on focal chroots
After hirsute/linux commit e3bd2870be7f "UBUNTU: [Packaging] replace
custom filter script with dctrl-tools", grep-dctrl is used to generate
the control files so the dctrl-tools package is needed. Creating the
hirsute chroot will pull this package automatically as it is a build
dependency, but for the backports in focal this needs to be explicitly
installed.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>