ancillary-bot: lrm-ancillary-lookup -- fix subscription handling
We have been hitting a but in lazr.restfulclient where an attempt
to lookup a private PPA subscription return the record for your
user rather than the requested URL.  We have worked around this
by scanning all of our subscriptions, but this cannot instantiate
a newly allocated one.  After much debugging this was tracked
to poor handling of the +me redirect which loses our POST request
making it a GET and thus getting our person object.  Identified
that if we lookup a user explicity the getArchiveSubscriptionURL
call then works.  Switched to getting our user from the +me object
and then looking ourselves up directly.

Signed-off-by: Andy Whitcroft <apw@canonical.com>