swm: wft/promote_to_proposed -- check both legs of a signing are clear
When we are splitting promote-to-proposed into two pieces to use a
signing PPA check that both the signing PPA and the destination are
clear before initiating the copy.  We do not want to block a kernel in
the signing PPA; this could block a security signing for instance.

Also new automation does not understand such things and must not be
triggered till both are clear.

Signed-off-by: Andy Whitcroft <apw@canonical.com>