swm: wft/promote_to_proposed -- consider p-t-p as a single step even for signing
Kernels can move from the build route to Proposed in three ways:

1) for unsigned kernels the are simply copied directly from build to
   Proposed,
2) a human firstly promotes them to Signing and once complete further
   promotes them to Proposed, and
3) canonical-signing-bot promotes the to Signing and once complete
   promotes them to Proposed.

In case (2) we have to consider the two steps independently as the
reviewer/promoter is driven by the promote-to-proposed and
promote-signing-to-proposed tasks.  In case (3) the bot does not
feedback to nor wait for the swm to be in sync with its actions, here we
need to consider only promote-to-proposed and have that encompass both
legs of the copy.

Differentiate cases (2) and (3) using a new tag kernel-signing-bot.
When present trigger closure of promote-signing-to-proposed to invalid
and change the meaning of promote-to-proposed.

Signed-off-by: Andy Whitcroft <apw@canonical.com>