swm: wft/promote_to_proposed -- monitor build progress
When scanning a pocket archive record monitor records for anything
we deem to be blocking to the packages being published in that
location.

When scanning a tracker in a pocket we will examine the source
publications, builds, uploads, and binary publications associated with
each of the dependant package associated with this tracker.  We record
the first item for each which is incomplete.  We perform the following
analysis for each dependant package in the set.

If the source publication record for a dependant package is either for the
wrong version or is not Published we will emit a launchpad-source record
allowing swm-publishing to determine when this changes (and trigger
scanning of the associated tracker).

As build record generation is non-atomic it is possible to scan a source
after publication but before its builds are recorded.  If the source
package is published but there are no builds we will emit a
launchpad-nobuilds record and swm-publishing will watch for the build
count becoming non-zero.

When we do have builds we examine each architecture build and for that
architecture record any which are not 'Successfully built'.  For each we
emit a launchpad-build record.  swm-publishing monitors these for any
change in status.  It also monitors the start time and also consider
this a change; this occurs when a human hits retry on a build in error
and it returns to that same state in the monitor period.

When builds are complete upload records are generated.  For any which
are not in Done or Rejected we emit a launchad-upload record.
swm-publishing monitors these for a change in status.

Once the uploads are compelete we scan the binary publication records
for the build.  If any are not Published we emit a launchpad-binary
record for the first such binary in the upload.  In the normal case
publication of all of the binaries in a particular upload are atomic so
monitoring one is a good proxy for the set.  swm-publishing monitors
these for a change in status.

We accumulate these records for any archive we scan for packages.  Later
when we know which if any are interesting to us we will emit the monitor
records for that subset of the archives.  For example if we are in the
middle of a promote-to-proposed we will monitor the destination pocket
for anything which is incomplete there.  We only monitor a pocket when
we consider we are actively building or promoting packages.  If
something was missed (such as a manual deletion of a package) we will
spot that at the next scheduled scan (at least every 4 hours).

Signed-off-by: Andy Whitcroft <apw@canonical.com>