swm: handle newly minted trackers in Triaged
When we instantiate a tracker it ends up in Triaged.  swm-publishing
will assume it needs scanning until it moves to In Progress.  As soon as
we have added it to the status file we should change the workflow task
In Progress.  This ensures it will be seen by swm-publishing and the bug
is not lost.

Signed-off-by: Andy Whitcroft <apw@canonical.com>