swm: wft/regression_testing -- send boot-testing messages earlier
Currently we send boot-testing messages as promote-to-proposed becomes
active.  Now that we have sru-review tasks this is delayed when that
review has not completed.  In order to allow boot-testing and sru-review
to progress in parallel we should request boot-testing when the builds
are complete.  Move requests to the boot-testing handler, emitting them
when we become active.  Only send them if we are not overridden already.

Signed-off-by: Andy Whitcroft <apw@canonical.com>