kernel-stable-dashboard: add early-testing status row and include boot-testing
Add an early-testing status row ("bite") similar to the general testing
row.  This will be used to summarise the parallel activities during
early build and testing in PPA.  Add the boot-testing status to this row
as the first entry.  The boot-testing result is already suppressed by
the existing *-testing suppression for the general testing row.

Signed-off-by: Andy Whitcroft <apw@canonical.com>