swm: swm-publishing -- monitor launchpad builds
Add monitor record handlers for the various launchpad-* record types.
In general this involved loading one launchpad object and comparing it
to the expected values.  Where they do not match a rescan is requested
for the associated tracker.  In each this is a single round-trip to
launchpad per launchpad-* monitor.

Now that we are monitoring builds via the monitor records we no longer
need to scan the monitor records for builds in progress as they are
effectivly duplicate.  By doing this we amortise a significant
proportion of the cost associated with launchpad-build monitor records
which are one of the longest living records.

Signed-off-by: Andy Whitcroft <apw@canonical.com>