swm: swm-publishing -- remove long redundant New bug tracking from MonitorStatusSwmStatus
We used to try and differentiate between a newly made tracker which is
tripping an exception and not writing a status and those which are
closing by looking at the status of the tracker in swm-publishing when
examining SwmStatus records.  We separatly are marking those using a new
kernel-sru-workflow task status so this code is redundant and long
commented.  Remove it and the factory components no longer used.

Signed-off-by: Andy Whitcroft <apw@canonical.com>