kernel-series: Change focal:linux-intel to be development
Right now there are still a bunch of changes expected to go directly
into the intel kernel (not via mailing list) and also it seems this
might need a quicker turn-around than we would be able to handle in
the stable cycles. For those reasons we change the definitions of
focal:linux-intel and focal:linux-uc20-intel to be development sources.
That way those are no longer included in stable cycles and tracking
bugs for those can be created in development cycle mode.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>