create-kernel-tasks: Optionally allow orphaned trackers
Add "--allow-nomaster" as an option which allows tracking bugs to be
created for sources which are derived from another source which does
not need an update but has no live or finalized tracking bug. This
seemed to have happened in some cases where we replaced a cycle with
security upodates which have no tracking bugs.

But the main use-case would be new sources which are derived from a
stable kernel but are still undergoing rapid changes and therefor do
not fit well into stable cycles.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>