swm: bug -- catch exceptions when loading swm-properties
If the bug swm-properties field fails to load we can end up losing the
data contained in it.  This leads the bug to become a "combo" tracker
and much of the key associated data is lost.  Detect these failures and
convert them into WorkflowCorruptErrors.  As we could not read the
swm-properties we cannot sync this into the bug without further damaging
the swm-properties, so we crowbar a crank-failure into the reasons field
in SwmInfo so that it will at least show up on the dashboard.

Signed-off-by: Andy Whitcroft <apw@canonical.com>